Move `NavMenu` to the `ui` module
