+++ WebRandomizer/ClientApp/src/ui/NavMenu.jsx
